+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Change Log
 HEAD
 ----
 * Added support for registering extended properties on folders.
+* Added support for creating, updating and deleting folders.
+
 
 1.10.6
 ------